added footer, new favicon, and reorganized components folders
