+++ src/App.css
@@ -12,75 +12,3 @@
     display: none !important;
   }
 }
-
-/* Navbar */
-
-.navbar {
-  display: flex;
-  justify-content: space-between;
-  width: 100%;
-  background-color: green;
-}
-
-.navbar .icon {
-  display: none;
-}
-
-.nav-name {
-  text-align: center;
-  margin: 0;
-  color: #FFF;
-  display: flex;
-  justify-content: space-between;
-  flex-direction: row;
-  flex-wrap: nowrap;
-}
-
-.nav-name a {
-  text-decoration: none;
-  color: #FFF;
-  padding: 25px;
-  margin: 10px;
-}
-
-.nav-name a:hover {
-  color: #99ccff;
-}
-
-.login-button {
-  height: 100%;
-}
-
-.login {
-  color: #FFF;
-}
-
-.login:hover {
-  color: #99ccff;
-}
-
-.button {
-  margin: 25px 25px 0px 0px !important;
-  height: 20% !important;
-  background-color: #1affff !important;
-}
-
-.button:hover {
-  opacity: 90% !important;
-}
-
-.logo-image {
-  width: 90px;
-  padding: 15px 0px 0px 25px;
-  border-radius: 50%;
-}
-
-.center-logo {
-  width: 150px;
-}
-
-.login-btn-group {
-  display: flex;
-  justify-content: space-between;
-}
-
