+++ src/App.js
@@ -1,6 +1,8 @@
+import { Navbar, Landing, Footer } from './components'
 import './App.css';
 import 'semantic-ui-css/semantic.min.css'
-import { Navbar, Landing } from './components'
+import 'antd/dist/antd.css';
+import 'font-awesome/css/font-awesome.min.css';
 
 function App() {
   return (
@@ -13,7 +15,7 @@ function App() {
         <Landing />
       </body>
       <footer>
-
+        <Footer />
       </footer>
     </div>
   );
