+++ src/components/footer/Footer.jsx
@@ -1,0 +1,12 @@
+import React from 'react';
+import './footer.css'
+
+const Footer = () => {
+    return (
+        <div>
+            <p>&copy; PetFinder 2021</p>
+        </div>
+    )
+}
+
+export default Footer;
