+++ src/components/footer/footer.css
@@ -1,0 +1,12 @@
+footer {
+    position: fixed;
+    left: 0;
+    bottom: 0;
+    width: 100%;
+    align-items: center;
+    background-color: green;
+    color: white;
+    display: flex;
+    justify-content: center;
+    padding: 30px 0;
+  }
