+++ src/components/index.js
@@ -1,4 +1,5 @@
-import Navbar from './Navbar';
+import Navbar from './navbar/Navbar';
 import Landing from './Landing';
+import Footer from './footer/Footer';
 
-export { Navbar, Landing };
+export { Navbar, Landing, Footer };
