+++ src/components/navbar/navbar.css
@@ -1,0 +1,95 @@
+/* Navbar */
+
+.navbar {
+    display: flex;
+    justify-content: space-between;
+    width: 100%;
+    background-color: green;
+  }
+  
+  .navbar .icon {
+    display: none;
+  }
+  
+  .nav-name {
+    text-align: center;
+    margin: 0;
+    color: #FFF;
+    display: flex;
+    justify-content: space-between;
+    flex-direction: row;
+    flex-wrap: nowrap;
+  }
+  
+  .nav-name a {
+    text-decoration: none;
+    color: #FFF;
+    padding: 25px;
+    margin: 10px;
+  }
+  
+  .nav-name a:hover {
+    color: #99ccff;
+  }
+  
+  .login-button {
+    height: 100%;
+    margin: 25px 25px 0px 0px;
+  }
+  
+  .login {
+    color: #FFF;
+  }
+  
+  .login:hover {
+    color: #99ccff;
+  }
+  
+  .button {
+    margin: 25px 25px 0px 0px !important;
+    height: 20% !important;
+    background-color: green !important;
+    color: #FFF !important;
+  }
+  
+  .fa-bars {
+    background-color: green;
+    border: none;
+    color: #FFF;
+    width: 100%;
+  }
+  
+  /* .primary {
+    background-color: green;
+    color: #FFF;
+    border: none;
+  }
+  
+  .primary:hover {
+    opacity: 90% !important;
+    color: aqua;
+  } */
+  
+  .logo-image {
+    width: 90px;
+    padding: 15px 0px 0px 25px;
+    border-radius: 50%;
+  }
+  
+  .center-logo {
+    width: 150px;
+  }
+  
+  .login-btn-group {
+    display: flex;
+    justify-content: space-between;
+  }
+  
+  .login-btn-group ul{
+    list-style: none;
+  }
+  
+  .login-btn-group ul li {
+    padding: 10px 0px 10px 0px;
+    width: 50%;
+  }
